% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmc.nsubs.moments}
\alias{ctmc.nsubs.moments}
\title{CTMC restricted moments of labeled substitution counts}
\usage{
ctmc.nsubs.moments(t, subst_mod, L, max_order)
}
\arguments{
\item{t}{A nonnegative numeric scalar representing the CTMC time interval 
length.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{L}{An indicator matrix that labels the substitutions of interest.}

\item{max_order}{An integer specifying the maximum moment order of interest.}
}
\value{
A three-dimensional array that stores the restricted moment 
  matrices, where the third dimension indexes the different moment orders.
}
\description{
Computes the CTMC restricted moment matrices of labeled substitution counts 
for all orders less than or equal to the user-specified maximum order of 
interest.
}
\details{
The zeroth-order restricted moment matrix is defined to be the transition 
probability matrix.
}
\references{
Van Loan C (1978) \dQuote{Computing Integrals Involving the 
  Matrix Exponential}, \emph{IEEE Transactions on Automatic Control}, 
  23(3):395-404.
  
  Minin VN and Suchard MA (2008) \dQuote{Counting labeled transitions in 
  continuous-time Markov models of evolution}, \emph{Journal of Mathematical
  Biology}, 56(3):391-412.
  
  OUR PAPER!!
}
\seealso{
\code{\link{ctmc.reward.moments}}, 
  \code{\link{phylo.nsubs.moments}}
}

