% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmc.Q.derivatives}
\alias{ctmc.Q.derivatives}
\title{CTMC transition probability derivatives with respect to rate matrix 
parameters}
\usage{
ctmc.Q.derivatives(t, subst_mod, param_name, max_order)
}
\arguments{
\item{t}{A nonnegative numeric scalar representing the CTMC time interval 
length.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{param_name}{The rate matrix parameter name of interest.}

\item{max_order}{An integer specifying the maximum derivative order of 
interest.}
}
\value{
A three-dimensional array that stores the transition probability 
  derivative matrices, where the third dimension indexes the different 
  derivative orders.  Specifically, if \code{outp} denotes the output array,
  then \code{outp[, , k]} represents the transition probability derivative 
  matrix of order \code{k - 1} for all integers \code{k} between \code{1} 
  and \code{max_order + 1}.
}
\description{
Computes the CTMC transition probability derivative matrices with respect to
a given rate matrix parameter for all orders less than or equal to the 
user-specified maximum order of interest.
}
\details{
The zeroth-order transition probability derivative matrix is defined to be 
the transition probability matrix.
}
\references{
Van Loan C (1978) \dQuote{Computing Integrals Involving the 
  Matrix Exponential}, \emph{IEEE Transactions on Automatic Control}, 
  23(3):395-404.
  
  Kenney T and Gu H (2012) \dQuote{Hessian Calculation for Phylogenetic 
  Likelihood based on the Pruning Algorithm and its Applications}, 
  \emph{Statistical Applications in Genetics and Molecular Biology}, 11(4).
  
  OUR PAPER!!
}
\seealso{
\code{\link{ctmc.t.derivatives}}, \code{\link{phylo.Q.derivatives}}
}

