% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmc.reward.moments}
\alias{ctmc.reward.moments}
\title{CTMC restricted moments of total rewards}
\usage{
ctmc.reward.moments(t, subst_mod, w, max_order)
}
\arguments{
\item{t}{A nonnegative numeric scalar representing the CTMC time interval 
length.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{w}{A numeric vector that defines the reward per unit time (reward 
rate) for each CTMC state.}

\item{max_order}{An integer specifying the maximum moment order of interest.}
}
\value{
A three-dimensional array that stores the restricted moment 
  matrices, where the third dimension indexes the different moment orders.
}
\description{
Computes the CTMC restricted moment matrices of total rewards for all orders
less than or equal to the user-specified maximum order of interest.
}
\details{
The zeroth-order restricted moment matrix is defined to be the transition 
probability matrix.
}
\references{
Van Loan C (1978) \dQuote{Computing Integrals Involving the 
  Matrix Exponential}, \emph{IEEE Transactions on Automatic Control}, 
  23(3):395-404.
  
  Minin VN and Suchard MA (2008) \dQuote{Fast, Accurate and Simulation-Free 
  Stochastic Mapping}, \emph{Philosophical Transactions of the Royal Society
  B: Biological Sciences}, 363(1512):3985-3995.
  
  OUR PAPER!!
}
\seealso{
\code{\link{ctmc.nsubs.moments}}, 
  \code{\link{phylo.reward.moments}}
}

