% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmc.sim}
\alias{ctmc.sim}
\title{CTMC path simulation}
\usage{
ctmc.sim(t, subst_mod, init_state)
}
\arguments{
\item{t}{A nonnegative numeric scalar representing the CTMC time interval 
length.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{init_state}{A string that specifies the initial state of the CTMC.}
}
\value{
A data frame with two columns.  The \code{"time"} column holds the 
  times at which each of the simulated CTMC states is entered and the 
  \code{"state"} column stores the corresponding CTMC states.
}
\description{
Simulates a CTMC sample path.
}
\references{
Nielsen R (2002) \dQuote{Mapping Mutations on Phylogenies}, 
  \emph{Systematic Biology}, 51(5):729-739.
}
\seealso{
\code{\link{asr.sim}}, \code{\link{smap.sim}}, 
  \code{\link{tips.sim}}
}

