% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmc.tpm}
\alias{ctmc.tpm}
\title{CTMC transition probability matrix}
\usage{
ctmc.tpm(t, subst_mod)
}
\arguments{
\item{t}{A nonnegative numeric scalar representing the CTMC time interval 
length.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}
}
\value{
A matrix that stores the CTMC transition probabilities.
}
\description{
Computes the CTMC transition probability matrix.
}
\seealso{
\code{\link{phylo.likelihood}}
}

