% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F81}
\alias{F81}
\title{The Felsenstein (F81) substitution model}
\usage{
F81(mu, pi, scale = FALSE)
}
\arguments{
\item{mu}{A nonnegative numeric scalar representing the overall substitution
rate.}

\item{pi}{A numeric vector defining the F81 stationary distribution.}

\item{scale}{A boolean indicating whether to scale the time dimension.  If
\code{TRUE}, then time is specified in terms of the expected number of
CTMC substitutions per site.}
}
\value{
A list (S3 object of class \code{"substitution.model"}) with the
  following named entries:
  \describe{
    \item{\code{states}}{The vector of DNA states.}
    \item{\code{Q}}{The F81 rate matrix.}
    \item{\code{pi}}{The F81 stationary distribution.}
    \item{\code{d_mu}}{The F81 rate matrix derivative with respect to
      \code{mu}.}
  }
}
\description{
Creates a Felsenstein (F81) substitution model object.
}
\references{
Felsenstein J (1981) \dQuote{Evolutionary Trees from DNA
  Sequences: A Maximum Likelihood Approach}, \emph{Journal of Molecular
  Evolution}, 17(6):368-376.
}
\seealso{
\code{\link{JC69}}, \code{\link{K80}}, \code{\link{HKY85}},
  \code{\link{GTR}}
}

