% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GTR}
\alias{GTR}
\title{The general time-reversible (GTR) substitution model}
\usage{
GTR(rAC, rAG, rAT, rCG, rCT, rGT, pi, scale = FALSE)
}
\arguments{
\item{rAC, rAG, rAT, rCG, rCT, rGT}{Nonnegative numeric scalars representing the
\eqn{A <=> C}, \eqn{A <=> G}, \eqn{A <=> T}, \eqn{C <=> G}, \eqn{C <=> T},
and \eqn{G <=> T} substitution rates, respectively.}

\item{pi}{A numeric vector defining the GTR stationary distribution.}

\item{scale}{A boolean indicating whether to scale the time dimension.  If
\code{TRUE}, then time is specified in terms of the expected number of
CTMC substitutions per site.}
}
\value{
A list (S3 object of class \code{"substitution.model"}) with the
  following named entries:
  \describe{
    \item{\code{states}}{The vector of DNA states.}
    \item{\code{Q}}{The GTR rate matrix.}
    \item{\code{pi}}{The GTR stationary distribution.}
    \item{\code{d_rAC}}{The GTR rate matrix derivative with respect to
      \code{rAC}.}
    \item{\code{d_rAG}}{The GTR rate matrix derivative with respect to
      \code{rAG}.}
    \item{\code{d_rAT}}{The GTR rate matrix derivative with respect to
      \code{rAT}.}
    \item{\code{d_rCG}}{The GTR rate matrix derivative with respect to
      \code{rCG}.}
    \item{\code{d_rCT}}{The GTR rate matrix derivative with respect to
      \code{rCT}.}
    \item{\code{d_rGT}}{The GTR rate matrix derivative with respect to
      \code{rGT}.}
  }
}
\description{
Creates a general time-reversible (GTR) substitution model object.
}
\references{
Tavare S (1986) \dQuote{Some Probabilistic and Statistical
  Problems in the Analysis of DNA Sequences}, \emph{Lectures on Mathematics
  in the Life Sciences}, 17(2):57-86.
}
\seealso{
\code{\link{JC69}}, \code{\link{K80}}, \code{\link{F81}},
  \code{\link{HKY85}}
}

