% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HKY85}
\alias{HKY85}
\title{The Hasegawa-Kishino-Yano (HKY85) substitution model}
\usage{
HKY85(alpha, beta, pi, scale = FALSE)
}
\arguments{
\item{alpha}{A nonnegative numeric scalar representing the substitution rate
of DNA transitions (\eqn{A <=> G} and \eqn{C <=> T}).}

\item{beta}{A nonnegative numeric scalar representing the substitution rate
of DNA transversions (\eqn{A <=> C}, \eqn{A <=> T}, \eqn{C <=> G}, and
\eqn{G <=> T}).}

\item{pi}{A numeric vector defining the HKY85 stationary distribution.}

\item{scale}{A boolean indicating whether to scale the time dimension.  If
\code{TRUE}, then time is specified in terms of the expected number of
CTMC substitutions per site.}
}
\value{
A list (S3 object of class \code{"substitution.model"}) with the
  following named entries:
  \describe{
    \item{\code{states}}{The vector of DNA states.}
    \item{\code{Q}}{The HKY85 rate matrix.}
    \item{\code{pi}}{The HKY85 stationary distribution.}
    \item{\code{d_alpha}}{The HKY85 rate matrix derivative with respect to
      \code{alpha}.}
    \item{\code{d_beta}}{The HKY85 rate matrix derivative with respect to
      \code{beta}.}
  }
}
\description{
Creates a Hasegawa-Kishino-Yano (HKY85) substitution model object.
}
\references{
Hasegawa M, Kishino H, and Yano TA (1985) \dQuote{Dating of the
  Human-Ape Splitting by a Molecular Clock of Mitochondrial DNA}, \emph{
  Journal of Molecular Evolution}, 22(2):160-174.
}
\seealso{
\code{\link{JC69}}, \code{\link{K80}}, \code{\link{F81}},
  \code{\link{GTR}}
}

