% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{JC69}
\alias{JC69}
\title{The Jukes-Cantor (JC69) substitution model}
\usage{
JC69(mu, scale = FALSE)
}
\arguments{
\item{mu}{A nonnegative numeric scalar representing the overall substitution
rate.}

\item{scale}{A boolean indicating whether to scale the time dimension.  If
\code{TRUE}, then time is specified in terms of the expected number of
CTMC substitutions per site.}
}
\value{
A list (S3 object of class \code{"substitution.model"}) with the
  following named entries:
  \describe{
    \item{\code{states}}{The vector of DNA states.}
    \item{\code{Q}}{The JC69 rate matrix.}
    \item{\code{pi}}{The JC69 stationary distribution.}
    \item{\code{d_mu}}{The JC69 rate matrix derivative with respect to
      \code{mu}.}
  }
}
\description{
Creates a Jukes-Cantor (JC69) substitution model object.
}
\references{
Jukes TH and Cantor CR (1969) \dQuote{Evolution of Protein
  Molecules}, \emph{Mammalian Protein Metabolism}, (3):21-132.
}
\seealso{
\code{\link{K80}}, \code{\link{F81}}, \code{\link{HKY85}},
  \code{\link{GTR}}
}

