% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{K80}
\alias{K80}
\title{The Kimura 2-parameter (K80) substitution model}
\usage{
K80(alpha, beta, scale = FALSE)
}
\arguments{
\item{alpha}{A nonnegative numeric scalar representing the substitution rate
of DNA transitions (\eqn{A <=> G} and \eqn{C <=> T}).}

\item{beta}{A nonnegative numeric scalar representing the substitution rate
of DNA transversions (\eqn{A <=> C}, \eqn{A <=> T}, \eqn{C <=> G}, and
\eqn{G <=> T}).}

\item{scale}{A boolean indicating whether to scale the time dimension.  If
\code{TRUE}, then time is specified in terms of the expected number of
CTMC substitutions per site.}
}
\value{
A list (S3 object of class \code{"substitution.model"}) with the
  following named entries:
  \describe{
    \item{\code{states}}{The vector of DNA states.}
    \item{\code{Q}}{The K80 rate matrix.}
    \item{\code{pi}}{The K80 stationary distribution.}
    \item{\code{d_alpha}}{The K80 rate matrix derivative with respect to
      \code{alpha}.}
    \item{\code{d_beta}}{The K80 rate matrix derivative with respect to
      \code{beta}.}
  }
}
\description{
Creates a Kimura 2-parameter (K80) substitution model object.
}
\references{
Kimura M (1980) \dQuote{A Simple Method for Estimating
  Evolutionary Rates of Base Substitutions Through Comparative Studies of
  Nucleotide Sequences}, \emph{Journal of Molecular Evolution},
  16(2):111-120.
}
\seealso{
\code{\link{JC69}}, \code{\link{F81}}, \code{\link{HKY85}},
  \code{\link{GTR}}
}

