% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phylo.likelihood}
\alias{phylo.likelihood}
\title{Phylogenetic likelihood}
\usage{
phylo.likelihood(tree, subst_mod, tip_states)
}
\arguments{
\item{tree}{An S3 object of class \code{"phylo"}.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{tip_states}{A character vector of observed tip states.}
}
\value{
A double-precision value that holds the phylogenetic likelihood.
}
\description{
Computes the phylogenetic likelihood.
}
\details{
More information on S3 objects of class \code{"phylo"} is found at 
\url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.

If an observed tip state is invalid, then it is treated as an ambiguous 
character state.
}
\references{
Felsenstein J (1981) \dQuote{Evolutionary Trees from DNA 
  Sequences: A Maximum Likelihood Approach}, \emph{Journal of Molecular 
  Evolution}, 17(6):368-376.
}
\seealso{
\code{\link{ctmc.tpm}}
}

