% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phylo.Q.derivatives}
\alias{phylo.Q.derivatives}
\title{Phylogenetic likelihood derivatives with respect to rate matrix parameters}
\usage{
phylo.Q.derivatives(tree, subst_mod, param_name, max_order, tip_states)
}
\arguments{
\item{tree}{An S3 object of class \code{"phylo"}.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{param_name}{The rate matrix parameter name of interest.}

\item{max_order}{An integer specifying the maximum derivative order of 
interest.}

\item{tip_states}{A character vector of observed tip states.}
}
\value{
A named numeric vector that holds the computed likelihood 
  derivatives.  Each element name is a string of the form \code{"(1):O"}, 
  where \code{O} represents the derivative order of the corresponding 
  element.  However, the empty string (\code{""}) is the element name of the
  zeroth-order likelihood derivative.
}
\description{
Computes the phylogenetic likelihood derivatives with respect to a given 
rate matrix parameter for all orders less than or equal to the 
user-specified maximum order of interest.
}
\details{
More information on S3 objects of class \code{"phylo"} is found at 
\url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.

The zeroth-order likelihood derivative is defined to be the phylogenetic 
likelihood.
}
\references{
Kenney T and Gu H (2012) \dQuote{Hessian Calculation for 
  Phylogenetic Likelihood based on the Pruning Algorithm and its 
  Applications}, \emph{Statistical Applications in Genetics and Molecular 
  Biology}, 11(4).
  
  OUR PAPER!!
}
\seealso{
\code{\link{phylo.t.derivatives}}, \code{\link{ctmc.Q.derivatives}}
}

