% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phylo.reward.moments}
\alias{phylo.reward.moments}
\title{Phylogenetic stochastic mapping moments of total rewards}
\usage{
phylo.reward.moments(tree, subst_mod, w, edge_sets, max_order, tip_states)
}
\arguments{
\item{tree}{An S3 object of class \code{"phylo"}.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{w}{A numeric vector that defines the reward per unit time (reward 
rate) for each CTMC state.}

\item{edge_sets}{A list of integer vectors that defines the edge sets over 
which the moments are calculated.}

\item{max_order}{An integer specifying the maximum moment order of interest.}

\item{tip_states}{A character vector of observed tip states.}
}
\value{
A named numeric vector that holds the computed stochastic mapping 
  moments.  Each element name is formed by concatenating strings of the form
  \code{"(E):O"} with \code{"--"} separators, where \code{E} denotes an edge
  set index, \code{O} represents the exponent order associated with the 
  additive mapping summary defined over the \code{E}th edge set, and 
  \code{"--"} symbolizes the product operator between two mapping summaries.
  However, the empty string (\code{""}) is the element name of the 
  zeroth-order stochastic mapping moment.
}
\description{
Computes the phylogenetic stochastic mapping moments of total rewards for 
all orders less than or equal to the user-specified maximum order of 
interest.
}
\details{
More information on S3 objects of class \code{"phylo"} is found at 
\url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.

The zeroth-order stochastic mapping moment is defined to be the phylogenetic
likelihood.

If the number of observed tip states (\code{length(tip_states)}) is not 
equal to the number of tips on the phylogeny 
(\code{length(tree$tip.label)}), then the stochastic mapping moments are 
computed using ambiguous character states.
}
\references{
Minin VN and Suchard MA (2008) \dQuote{Fast, Accurate and 
  Simulation-Free Stochastic Mapping}, \emph{Philosophical Transactions of 
  the Royal Society B: Biological Sciences}, 363(1512):3985-3995.
  
  Dhar A and Minin VN (2017) \dQuote{Calculating Higher-Order Moments of 
  Phylogenetic Stochastic Mapping Summaries in Linear Time}, \emph{Journal 
  of Computational Biology}, 24(5):377-399.
  
  OUR PAPER!!
}
\seealso{
\code{\link{phylo.nsubs.moments}}, 
  \code{\link{ctmc.reward.moments}}
}

