% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phylo.t.derivatives}
\alias{phylo.t.derivatives}
\title{Phylogenetic likelihood derivatives with respect to branch lengths}
\usage{
phylo.t.derivatives(tree, subst_mod, max_order, tip_states)
}
\arguments{
\item{tree}{An S3 object of class \code{"phylo"}.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{max_order}{An integer specifying the maximum derivative order of 
interest.}

\item{tip_states}{A character vector of observed tip states.}
}
\value{
A named numeric vector that holds the computed likelihood 
  derivatives.  Each element name is formed by concatenating strings of the 
  form \code{"(E):O"} with \code{"--"} separators, where \code{E} denotes an
  edge index, \code{O} represents the derivative order associated with the 
  edge index \code{E}, and \code{"--"} symbolizes the partial 
  differentiation operator with respect to two branch lengths.  However, the
  empty string (\code{""}) is the element name of the zeroth-order 
  likelihood derivative.
}
\description{
Computes the phylogenetic likelihood derivatives with respect to the branch 
lengths for all orders less than or equal to the user-specified maximum 
order of interest.
}
\details{
More information on S3 objects of class \code{"phylo"} is found at 
\url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.

The zeroth-order likelihood derivative is defined to be the phylogenetic 
likelihood.

If an observed tip state is invalid, then it is treated as an ambiguous 
character state.
}
\references{
Kenney T and Gu H (2012) \dQuote{Hessian Calculation for 
  Phylogenetic Likelihood based on the Pruning Algorithm and its 
  Applications}, \emph{Statistical Applications in Genetics and Molecular 
  Biology}, 11(4).
  
  OUR PAPER!!
}
\seealso{
\code{\link{phylo.Q.derivatives}}, \code{\link{ctmc.t.derivatives}}
}

