% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomd.R
\docType{package}
\name{phylomd}
\alias{phylomd}
\alias{phylomd-package}
\title{Efficient Computations of Higher-Order Phylogenetic Moments and Derivatives}
\description{
This package implements a post-order tree traversal meta-algorithm that 
computes phylogenetic stochastic mapping moments and likelihood derivatives 
of any order and scales linearly in the number of phylogeny tips.  Utility 
functions related to simulation-based stochastic mapping and higher-order 
continuous-time Markov chain (CTMC) moment and derivative calculations are 
also provided.
}
\author{
Amrit Dhar, Vladimir N. Minin
  
  Maintainer: Amrit Dhar \email{adhar@uw.edu}
}
\references{
Nielsen R (2002) \dQuote{Mapping Mutations on Phylogenies}, 
  \emph{Systematic Biology}, 51(5):729-739.
  
  Minin VN and Suchard MA (2008) \dQuote{Counting labeled transitions in 
  continuous-time Markov models of evolution}, \emph{Journal of Mathematical 
  Biology}, 56(3):391-412.
  
  Minin VN and Suchard MA (2008) \dQuote{Fast, Accurate and Simulation-Free 
  Stochastic Mapping}, \emph{Philosophical Transactions of the Royal Society 
  B: Biological Sciences}, 363(1512):3985-3995.
  
  Kenney T and Gu H (2012) \dQuote{Hessian Calculation for Phylogenetic 
  Likelihood based on the Pruning Algorithm and its Applications}, 
  \emph{Statistical Applications in Genetics and Molecular Biology}, 11(4).
  
  Dhar A and Minin VN (2017) \dQuote{Calculating Higher-Order Moments of 
  Phylogenetic Stochastic Mapping Summaries in Linear Time}, \emph{Journal of
  Computational Biology}, 24(5):377-399.
  
  OUR PAPER!!
}

