% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smap.sim}
\alias{smap.sim}
\title{Phylogenetic stochastic mapping simulation}
\usage{
smap.sim(tree, subst_mod, tip_states)
}
\arguments{
\item{tree}{An S3 object of class \code{"phylo"}.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}

\item{tip_states}{A character vector of observed tip states.}
}
\value{
A list that stores the simulated CTMC trajectory at each edge.
}
\description{
Samples a stochastic mapping on a phylogeny conditional on observed tip 
states.
}
\details{
More information on S3 objects of class \code{"phylo"} is found at 
\url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\references{
Nielsen R (2002) \dQuote{Mapping Mutations on Phylogenies}, 
  \emph{Systematic Biology}, 51(5):729-739.
}
\seealso{
\code{\link{asr.sim}}, \code{\link{tips.sim}}, 
  \code{\link{ctmc.sim}}
}

