% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tips.sim}
\alias{tips.sim}
\title{Phylogenetic tip state simulation}
\usage{
tips.sim(tree, subst_mod)
}
\arguments{
\item{tree}{An S3 object of class \code{"phylo"}.}

\item{subst_mod}{An S3 object of class \code{"substitution.model"}.}
}
\value{
A character vector of simulated tip states.
}
\description{
Simulates the tip states of a phylogeny.
}
\details{
More information on S3 objects of class \code{"phylo"} is found at 
\url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\seealso{
\code{\link{asr.sim}}, \code{\link{smap.sim}}, 
  \code{\link{ctmc.sim}}
}

